refactor: move dateformat_tags to apps/web/templatetags
Update all import sites (models.py, tasks.py, test file) to use the
new location. Django template {% load %} directives are unaffected as
they load by module name.
